Resolve header file conflict between Qt and lib3mf.
